+++ migrations/015_verification.sql
@@ -8,7 +8,7 @@
 
 -- ── buyer_contacts ────────────────────────────────────────────────────────────
 ALTER TABLE buyer_contacts ADD COLUMN IF NOT EXISTS is_verified INTEGER DEFAULT 0;
-ALTER TABLE buyer_contacts ADD COLUMN IF NOT EXISTS verified_at INTEGER;
+ALTER TABLE buyer_contacts ADD COLUMN IF NOT EXISTS verified_at BIGINT;
 ALTER TABLE buyer_contacts ADD COLUMN IF NOT EXISTS verified_by TEXT;
 
 UPDATE buyer_contacts
@@ -19,7 +19,7 @@ WHERE is_verified = 0;
 
 -- ── production_companies ──────────────────────────────────────────────────────
 ALTER TABLE production_companies ADD COLUMN IF NOT EXISTS is_verified INTEGER DEFAULT 0;
-ALTER TABLE production_companies ADD COLUMN IF NOT EXISTS verified_at INTEGER;
+ALTER TABLE production_companies ADD COLUMN IF NOT EXISTS verified_at BIGINT;
 ALTER TABLE production_companies ADD COLUMN IF NOT EXISTS verified_by TEXT;
 
 UPDATE production_companies
@@ -30,7 +30,7 @@ WHERE is_verified = 0;
 
 -- ── shows ─────────────────────────────────────────────────────────────────────
 ALTER TABLE shows ADD COLUMN IF NOT EXISTS is_verified INTEGER DEFAULT 0;
-ALTER TABLE shows ADD COLUMN IF NOT EXISTS verified_at INTEGER;
+ALTER TABLE shows ADD COLUMN IF NOT EXISTS verified_at BIGINT;
 ALTER TABLE shows ADD COLUMN IF NOT EXISTS verified_by TEXT;
 
 -- Only verify non-trade shows: 'manual' and 'sheet' data_source rows are trusted.
@@ -42,7 +42,7 @@ WHERE data_source IN ('manual', 'sheet') OR data_source IS NULL;
 
 -- ── mandate_updates ───────────────────────────────────────────────────────────
 ALTER TABLE mandate_updates ADD COLUMN IF NOT EXISTS is_verified INTEGER DEFAULT 0;
-ALTER TABLE mandate_updates ADD COLUMN IF NOT EXISTS verified_at INTEGER;
+ALTER TABLE mandate_updates ADD COLUMN IF NOT EXISTS verified_at BIGINT;
 ALTER TABLE mandate_updates ADD COLUMN IF NOT EXISTS verified_by TEXT;
 
 UPDATE mandate_updates
@@ -53,6 +53,6 @@ WHERE source IS NULL OR source NOT IN ('trade', 'scraper');
 
 -- ── market_orders ─────────────────────────────────────────────────────────────
 ALTER TABLE market_orders ADD COLUMN IF NOT EXISTS is_verified INTEGER DEFAULT 0;
-ALTER TABLE market_orders ADD COLUMN IF NOT EXISTS verified_at INTEGER;
+ALTER TABLE market_orders ADD COLUMN IF NOT EXISTS verified_at BIGINT;
 ALTER TABLE market_orders ADD COLUMN IF NOT EXISTS verified_by TEXT;
 -- All market_orders come from scrapers — leave is_verified=0 (no UPDATE).
